+++ 2024/libs/src/array_list.c
@@ -19,7 +19,7 @@ ArrayList *array_list_init(const uint32_t capacity) {
   return list;
 }
 
-void array_list_add(ArrayList *this, uint16_t value) {
+void array_list_add(ArrayList *this, uint32_t value) {
   if (this->count == this->capacity) {
     printf("Array list is full!\n");
     return;
@@ -27,7 +27,7 @@ void array_list_add(ArrayList *this, uint16_t value) {
   this->head[this->count++] = value;
 }
 
-uint32_t array_list_get(const ArrayList *this, const uint16_t index) {
+uint32_t array_list_get(const ArrayList *this, const uint32_t index) {
   if (index > this->count) {
     printf("Invalid array list index\n");
     return 0;
@@ -35,8 +35,8 @@ uint32_t array_list_get(const ArrayList *this, const uint16_t index) {
   return this->head[index];
 }
 
-bool array_list_contains(const ArrayList *this, const uint16_t value) {
-  for (uint16_t i = 0; i < this->count; ++i) {
+bool array_list_contains(const ArrayList *this, const uint32_t value) {
+  for (uint32_t i = 0; i < this->count; ++i) {
     if (this->head[i] == value) {
       return true;
     }
